requirements.txt
torch>=1.10.0
transformers>=4.21.0
datasets>=2.6.1
peft>=0.3.0
bitsandbytes>=0.37.0
accelerate>=0.18.0
wandb>=0.13.5
boto3>=1.26.0
smart_open>=6.3.0
python-dotenv>=0.21.0
numpy>=1.21.0
tqdm>=4.64.0
